Add some translations to pk-debuginfo-install, and generally make more sane

+++ contrib/debuginfo-install/pk-debuginfo-install.c
@@ -19,6 +19,8 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  */
 
+/* Test with pk-debuginfo-install bzip2-libs-1.0.5-5.fc11.i586 glib2-2.20.3-1.fc11.i586 */
+
 #include "config.h"
 
 #include <string.h>
@@ -27,6 +29,12 @@
 #include <packagekit-glib/packagekit.h>
 
 #include "egg-debug.h"
+
+typedef struct {
+	GPtrArray *enabled;
+	GPtrArray *disabled;
+	PkClient *client;
+} PkDebuginfoInstallPrivate;
 
 /**
  * pk_get_package_name_from_nevra:
@@ -63,17 +71,11 @@ out:
 	return name;
 }
 
-typedef struct {
-	GPtrArray *enabled;
-	GPtrArray *disabled;
-	PkClient *client;
-} PkPriv;
-
 /**
- * repo_details_cb:
+ * pk_debuginfo_install_repo_details_cb:
  **/
 static void
-repo_details_cb (PkClient *client, const gchar *repo_id, const gchar *description, gboolean enabled, PkPriv *priv)
+pk_debuginfo_install_repo_details_cb (PkClient *client, const gchar *repo_id, const gchar *description, gboolean enabled, PkDebuginfoInstallPrivate *priv)
 {
 	if (enabled)
 		g_ptr_array_add (priv->enabled, g_strdup (repo_id));
@@ -82,10 +84,10 @@ repo_details_cb (PkClient *client, const gchar *repo_id, const gchar *descriptio
 }
 
 /**
- * is_in_array:
+ * pk_debuginfo_install_in_array:
  **/
 static gboolean
-is_in_array (GPtrArray *array, const gchar *text)
+pk_debuginfo_install_in_array (GPtrArray *array, const gchar *text)
 {
 	guint i;
 	gboolean ret = FALSE;
@@ -103,10 +105,10 @@ is_in_array (GPtrArray *array, const gchar *text)
 }
 
 /**
- * enable_repos:
+ * pk_debuginfo_install_enable_repos:
  **/
 static gboolean
-enable_repos (PkPriv *priv, GPtrArray *array, gboolean enable, GError **error)
+pk_debuginfo_install_enable_repos (PkDebuginfoInstallPrivate *priv, GPtrArray *array, gboolean enable, GError **error)
 {
 	guint i;
 	gboolean ret = TRUE;
@@ -139,10 +141,10 @@ out:
 }
 
 /**
- * install_packages:
+ * pk_debuginfo_install_packages_install:
  **/
 static gboolean
-install_packages (PkPriv *priv, GPtrArray *array, GError **error)
+pk_debuginfo_install_packages_install (PkDebuginfoInstallPrivate *priv, GPtrArray *array, GError **error)
 {
 	gboolean ret = TRUE;
 	gchar **package_ids;
@@ -172,10 +174,10 @@ out:
 }
 
 /**
- * resolve_to_id:
+ * pk_debuginfo_install_resolve_name_to_id:
  **/
 static gchar *
-resolve_to_id (PkPriv *priv, const gchar *package_name, GError **error)
+pk_debuginfo_install_resolve_name_to_id (PkDebuginfoInstallPrivate *priv, const gchar *package_name, GError **error)
 {
 	gboolean ret;
 	const PkPackageObj *obj;
@@ -229,6 +231,51 @@ out:
 }
 
 /**
+ * pk_debuginfo_install_remove_suffix:
+ **/
+static gboolean
+pk_debuginfo_install_remove_suffix (gchar *name, const gchar *suffix)
+{
+	gboolean ret = FALSE;
+	guint slen, len;
+
+	if (!g_str_has_suffix (name, suffix))
+		goto out;
+
+	/* get lengths */
+	len = strlen (name);
+	slen = strlen (suffix);
+
+	/* same string */
+	if (len == slen)
+		goto out;
+
+	/* truncate */
+	name[len-slen] = '\0';
+	ret = TRUE;
+out:
+	return ret;
+}
+
+/**
+ * pk_debuginfo_install_print_array:
+ **/
+static void
+pk_debuginfo_install_print_array (GPtrArray *array)
+{
+	guint i;
+	PkPackageId *id;
+	const gchar *package_id;
+
+	for (i=0; i<array->len; i++) {
+		package_id = g_ptr_array_index (array, i);
+		id = pk_package_id_new_from_string (package_id);
+		g_print ("%i\t%s-%s(%s)\t%s\n", i, id->name, id->version, id->arch, id->data);
+		pk_package_id_free (id);
+	}
+}
+
+/**
  * main:
  **/
 int
@@ -243,19 +290,25 @@ main (int argc, char *argv[])
 	gchar *name;
 	gchar *name_debuginfo;
 	gboolean verbose = FALSE;
+	gboolean simulate = FALSE;
 	GOptionContext *context;
 	const gchar *repo_id;
 	gchar *repo_id_debuginfo;
-	PkPriv _priv;
-	PkPriv *priv = &_priv;
+	PkDebuginfoInstallPrivate _priv;
+	PkDebuginfoInstallPrivate *priv = &_priv;
+	guint step = 1;
 
 	const GOptionEntry options[] = {
 		{ "verbose", 'v', 0, G_OPTION_ARG_NONE, &verbose,
 		  _("Show extra debugging information"), NULL },
+		{ "simulate", 's', 0, G_OPTION_ARG_NONE, &simulate,
+		   /* command line argument, simulate what would be done, but don't actually do it */
+		  _("Don't actually install any packages, only simulate"), NULL },
 		{ NULL}
 	};
 
-	memset (priv, 0, sizeof (PkPriv));
+	/* clear private struct */
+	memset (priv, 0, sizeof (PkDebuginfoInstallPrivate));
 
 	setlocale (LC_ALL, "");
 	bindtextdomain (GETTEXT_PACKAGE, PACKAGE_LOCALE_DIR);
@@ -268,7 +321,7 @@ main (int argc, char *argv[])
 
 	context = g_option_context_new (NULL);
 	/* TRANSLATORS: tool that gets called when the command is not found */
-	g_option_context_set_summary (context, _("PackageKit Command Not Found"));
+	g_option_context_set_summary (context, _("PackageKit Debuginfo Installer"));
 	g_option_context_add_main_entries (context, options, NULL);
 	g_option_context_parse (context, &argc, &argv, NULL);
 	g_option_context_free (context);
@@ -277,7 +330,9 @@ main (int argc, char *argv[])
 
 	/* no input */
 	if (argv[1] == NULL) {
-		g_print ("specify packages to install\n");
+		/* TRANSLATORS: the use needs to specify a list of package names on the command line */
+		g_print (_("ERROR: Specify package names to install."));
+		g_print ("\n");
 		goto out;
 	}
 
@@ -289,22 +344,38 @@ main (int argc, char *argv[])
 
 	/* create #PkClient */
 	priv->client = pk_client_new ();
-	g_signal_connect (priv->client, "repo-detail", G_CALLBACK (repo_details_cb), priv);
+	g_signal_connect (priv->client, "repo-detail", G_CALLBACK (pk_debuginfo_install_repo_details_cb), priv);
 	pk_client_set_synchronous (priv->client, TRUE, NULL);
 	pk_client_set_use_buffer (priv->client, TRUE, NULL);
 
+	/* starting this section */
+	g_print ("%i. ", step++);
+
+	/* TRANSLATORS: we are getting the list of repositories */
+	g_print (_("Getting sources list..."));
+
 	/* get all enabled repos */
-	g_print ("1. Getting sources list...");
 	ret = pk_client_get_repo_list (priv->client, PK_FILTER_ENUM_NONE, &error);
 	if (!ret) {
 		g_print ("failed to get repo list: %s", error->message);
 		g_error_free (error);
 		goto out;
 	}
-	g_print ("OK. Found %i enabled and %i disabled sources.\n", priv->enabled->len, priv->disabled->len);
+
+	/* TRANSLATORS: all completed 100% */
+	g_print ("%s ", _("OK."));
+
+	/* TRANSLATORS: tell the user what we found */
+	g_print (_("Found %i enabled and %i disabled sources."), priv->enabled->len, priv->disabled->len);
+	g_print ("\n");
+
+	/* starting this section */
+	g_print ("%i. ", step++);
+
+	/* TRANSLATORS: we're finding repositories that match out pattern */
+	g_print (_("Finding debugging sources..."));
 
 	/* find all debuginfo repos for repos that are enabled */
-	g_print ("2. Finding debug sources...");
 	for (i=0; i<priv->enabled->len; i++) {
 
 		/* is already a -debuginfo */
@@ -316,7 +387,7 @@ main (int argc, char *argv[])
 
 		/* has a debuginfo repo */
 		repo_id_debuginfo = g_strjoin ("-", repo_id, "debuginfo", NULL);
-		ret = is_in_array (priv->disabled, repo_id_debuginfo);
+		ret = pk_debuginfo_install_in_array (priv->disabled, repo_id_debuginfo);
 		if (ret) {
 			/* add to list to change back at the end */
 			g_ptr_array_add (added_repos, g_strdup (repo_id_debuginfo));
@@ -326,31 +397,62 @@ main (int argc, char *argv[])
 
 		g_free (repo_id_debuginfo);
 	}
-	g_print ("OK. Found %i non-enabled debuginfo repos.\n", added_repos->len);
+
+	/* TRANSLATORS: all completed 100% */
+	g_print ("%s ", _("OK."));
+
+	/* TRANSLATORS: tell the user what we found */
+	g_print (_("Found %i disabled debuginfo repos."), added_repos->len);
+	g_print ("\n");
+
+	/* starting this section */
+	g_print ("%i. ", step++);
+
+	/* TRANSLATORS: we're now enabling all the debug sources we found */
+	g_print (_("Enabling debugging sources..."));
 
 	/* enable all debuginfo repos we found */
-	g_print ("3. Enabling debug sources...");
-	ret = enable_repos (priv, added_repos, TRUE, &error);
+	ret = pk_debuginfo_install_enable_repos (priv, added_repos, TRUE, &error);
 	if (!ret) {
-		g_print ("failed to enable repos: %s", error->message);
+		/* TRANSLATORS: operation was not successful */
+		g_print ("%s ", _("FAILED."));
+
+		/* TRANSLATORS: we're failed to enable the sources, detailed error follows */
+		g_print ("Failed to enable debugging sources: %s", error->message);
+		g_print ("\n");
 		g_error_free (error);
 		goto out;
 	}
-	g_print ("OK. Enabled %i sources.\n", added_repos->len);
 
-//debuginfo-install bzip2-libs-1.0.5-5.fc11.i586 glib2-2.20.3-1.fc11.i586 libacl-2.2.47-4.fc11.i586 libarchive-2.6.2-1.fc11.i586 lzma-libs-4.32.7-2.fc11.i586 sqlite-3.6.12-3.fc11.i586 zlib-1.2.3-22.fc11.i586
+	/* TRANSLATORS: all completed 100% */
+	g_print ("%s ", _("OK."));
+
+	/* TRANSLATORS: tell the user how many we enabled */
+	g_print (_("Enabled %i debugging sources."), added_repos->len);
+	g_print ("\n");
+
+	/* starting this section */
+	g_print ("%i. ", step++);
+
+	/* TRANSLATORS: we're now finding packages that match in all the repos */
+	g_print ("Finding debugging packages...");
 
 	/* parse arguments and resolve to packages */
-	g_print ("4. Finding debug packages...");
 	for (i=1; argv[i] != NULL; i++) {
 		name = pk_get_package_name_from_nevra (argv[i]);
+
+		/* remove suffix */
+		pk_debuginfo_install_remove_suffix (name, "-libs");
+
 		name_debuginfo = g_strjoin ("-", name, "debuginfo", NULL);
 		egg_debug ("install %s [%s]", argv[i], name_debuginfo);
 
 		/* resolve name */
-		package_id = resolve_to_id (priv, name_debuginfo, &error);
+		package_id = pk_debuginfo_install_resolve_name_to_id (priv, name_debuginfo, &error);
 		if (package_id == NULL) {
-			g_print ("failed to map package name to an id: %s\n", error->message);
+			/* TRANSLATORS: we couldn't find the package name, non-fatal */
+			g_print (_("Failed to find a package of that name: %s"), error->message);
+			g_print ("\n");
 			g_error_free (error);
 			/* don't quit, this is non-fatal */
 			error = NULL;
@@ -358,7 +460,7 @@ main (int argc, char *argv[])
 
 		/* add to array to install */
 		if (package_id != NULL) {
-			egg_debug ("going to try to install: %s\n", package_id);
+			egg_debug ("going to try to install: %s", package_id);
 			g_ptr_array_add (package_ids_to_install, g_strdup (package_id));
 		}
 
@@ -367,18 +469,51 @@ main (int argc, char *argv[])
 		g_free (package_id);
 
 	}
-	g_print ("OK. Found %i packages.\n", package_ids_to_install->len);
+	if (package_ids_to_install->len == 0) {
+		/* TRANSLATORS: operation was not successful */
+		g_print ("%s ", _("FAILED."));
+
+		/* TRANSLATORS: no debuginfo packages could be found to be installed */
+		g_print (_("Found no packages to install."));
+		g_print ("\n");
+		goto out;
+	}
+
+	/* TRANSLATORS: all completed 100% */
+	g_print ("%s ", _("OK."));
+
+	/* TRANSLATORS: tell the user we found some packages, and then list them */
+	g_print (_("Found %i packages:"), package_ids_to_install->len);
+	g_print ("\n");
+
+	/* print list */
+	pk_debuginfo_install_print_array (package_ids_to_install);
+
+	/* simulate mode for testing */
+	if (simulate) {
+		/* TRANSLATORS: simulate mode is a testing mode where we quit before the action */
+		g_print (_("Not installing packages in simulate mode"));
+		g_print ("\n");
+		goto out;
+	}
+
+	/* starting this section */
+	g_print ("%i. ", step++);
+
+	/* TRANSLATORS: we are now installing the debuginfo packages we found earlier */
+	g_print ("Installing packages...");
 
 	/* install */
-	g_print ("5. Installing packages...");
-	ret = install_packages (priv, package_ids_to_install, &error);
+	ret = pk_debuginfo_install_packages_install (priv, package_ids_to_install, &error);
 	if (!ret) {
 		g_print ("failed to install packages: %s", error->message);
 		g_error_free (error);
 		goto out;
 	}
-	g_print ("OK.\n");
 
+	/* TRANSLATORS: all completed 100% */
+	g_print (_("OK."));
+	g_print ("\n");
 out:
 
 	if (package_ids_to_install != NULL) {
@@ -387,14 +522,31 @@ out:
 	}
 	if (added_repos != NULL) {
 
-		g_print ("6. Disabling sources...");
+		/* starting this section */
+		g_print ("%i. ", step++);
+
+		/* TRANSLATORS: we are now disabling all debuginfo repos we previously enabled */
+		g_print (_("Disabling sources previously enabled..."));
+
 		/* disable all debuginfo repos we previously enabled */
-		ret = enable_repos (priv, added_repos, FALSE, &error);
+		ret = pk_debuginfo_install_enable_repos (priv, added_repos, FALSE, &error);
 		if (!ret) {
-			g_print ("failed to enable repos: %s", error->message);
+			/* TRANSLATORS: operation was not successful */
+			g_print ("%s ", _("FAILED."));
+
+			/* TRANSLATORS: no debuginfo packages could be found to be installed, detailed error follows */
+			g_print (_("Could not disable the debugging sources: %s"), error->message);
+			g_print ("\n");
 			g_error_free (error);
+		} else {
+
+			/* TRANSLATORS: all completed 100% */
+			g_print ("%s ", _("OK."));
+
+			/* TRANSLATORS: we disabled all the debugging repos that we enabled before */
+			g_print (_("Disabled %i debugging sources."), added_repos->len);
+			g_print ("\n");
 		}
-		g_print ("OK. Disabled %i sources.\n", added_repos->len);
 
 		g_ptr_array_foreach (added_repos, (GFunc) g_free, NULL);
 		g_ptr_array_free (added_repos, TRUE);
